+++ lib/vcr/util/hooks.rb
@@ -2,6 +2,10 @@ module VCR
   module Hooks
     include VariableArgsBlockCaller
 
+    def self.included(klass)
+      klass.extend(ClassMethods)
+    end
+
     def invoke_hook(hook, tag, *args)
       hooks_for(hook, tag).each do |callback|
         call_block(callback, *args)
@@ -12,16 +16,24 @@ module VCR
       hooks.clear
     end
 
-    private
+  private
 
-      def hooks
-        @hooks ||= Hash.new do |hook_type_hash, hook_type|
-          hook_type_hash[hook_type] = Hash.new do |tag_hash, tag|
-            tag_hash[tag] = []
-          end
+    def hooks
+      @hooks ||= Hash.new do |hook_type_hash, hook_type|
+        hook_type_hash[hook_type] = Hash.new do |tag_hash, tag|
+          tag_hash[tag] = []
         end
       end
+    end
 
+    def hooks_for(hook, tag)
+      for_hook = hooks[hook]
+      hooks = for_hook[tag] # matching tagged hooks
+      hooks += for_hook[nil] unless tag.nil? # untagged hooks
+      hooks
+    end
+
+    module ClassMethods
       def define_hook(hook)
         # We use splat args here because 1.8.7 doesn't allow default
         # values for block arguments, so we have to fake it.
@@ -34,12 +46,6 @@ module VCR
           hooks[hook][tag] << block
         end
       end
-
-      def hooks_for(hook, tag)
-        for_hook = hooks[hook]
-        hooks = for_hook[tag] # matching tagged hooks
-        hooks += for_hook[nil] unless tag.nil? # untagged hooks
-        hooks
-      end
+    end
   end
 end